Fix for Martingale+Predictor Paid).py
As reported by 'applez#0001:
"when i open the martingale+predictor, it just opens google and enters bet amounts and numbers but never actually presses the bet button"

Fix was fairly simple, I copied over the selection method fix that I did in the free version and used that to select the game button instead. There was no error message printing out and so I also added that in case there were any errors.